refactor(itinerary-body): Extract StayOnBoard component for others to reuse.
