+++ packages/itinerary-body/src/defaults/index.ts
@@ -4,6 +4,7 @@ import LineColumnContent from "./line-column-content";
 import PlaceName from "./place-name";
 import RouteDescription from "./route-description";
 import RouteLongName from "./route-long-name";
+import StayOnBoard from "./stay-on-board";
 import TimeColumnContent from "./time-column-content";
 import TransitLegSummary from "./transit-leg-summary";
 
@@ -14,6 +15,7 @@ export {
   PlaceName,
   RouteDescription,
   RouteLongName,
+  StayOnBoard,
   TimeColumnContent,
   TransitLegSummary
 };
